Decouple the api/server/router packages from the daemon/cluster and libnetwork packages.
