+++ api/server/backend/network.go
@@ -1,0 +1,224 @@
+package backend
+
+import (
+	"github.com/docker/engine-api/types"
+	"github.com/docker/engine-api/types/network"
+	"github.com/docker/libnetwork"
+)
+
+// GetNetworks returns a list of all networks.
+func (b *Backend) GetNetworks() ([]types.NetworkResource, error) {
+	list := []types.NetworkResource{}
+
+	if b.clusterProvider != nil {
+		if nr, err := b.clusterProvider.GetNetworks(); err == nil {
+			for _, nw := range nr {
+				list = append(list, nw)
+			}
+		}
+	}
+
+	// Combine the network list returned by Docker daemon if it is not already
+	// returned by the cluster manager
+SKIP:
+	for _, nw := range b.Daemon.GetNetworks() {
+		for _, nl := range list {
+			if nl.ID == nw.ID() {
+				continue SKIP
+			}
+		}
+		list = append(list, *b.buildNetworkResource(nw))
+	}
+
+	return list, nil
+}
+
+// FindNetwork function finds a network for a given string that can represent a network name or id.
+func (b *Backend) FindNetwork(idName string) (*types.NetworkResource, error) {
+	var err error
+	if nw, err := b.Daemon.FindNetwork(idName); err == nil {
+		return b.buildNetworkResource(nw), nil
+	}
+	if nr, err := b.clusterProvider.GetNetwork(idName); err == nil {
+		return &nr, nil
+	}
+	return nil, err
+}
+
+// CreateNetwork creates a network with the given name, driver and other optional parameters.
+func (b *Backend) CreateNetwork(create types.NetworkCreateRequest) (*types.NetworkCreateResponse, error) {
+	nw, err := b.Daemon.CreateNetwork(create)
+	if err == nil {
+		return nw, nil
+	}
+
+	// Ignore errors of type libnetwork.ManagerRedirectError
+	if _, ok := err.(libnetwork.ManagerRedirectError); !ok {
+		return nil, err
+	}
+
+	id, err := b.clusterProvider.CreateNetwork(create)
+	if err == nil {
+		return &types.NetworkCreateResponse{ID: id}, nil
+	}
+
+	return nil, err
+}
+
+// ConnectContainerToNetwork connects the given container to the
+// given network. If either cannot be found, an err is returned.
+// If the network cannot be set up, an err is returned.
+func (b *Backend) ConnectContainerToNetwork(containerID string, networkID string, endpointConfig *network.EndpointSettings) error {
+	nw, err := b.Daemon.FindNetwork(networkID)
+	if err != nil {
+		return err
+	}
+	return b.Daemon.ConnectContainerToNetwork(containerID, nw.Name(), endpointConfig)
+}
+
+// DisconnectContainerFromNetwork disconnects the given container from
+// the given network. If either cannot be found, an err is returned.
+func (b *Backend) DisconnectContainerFromNetwork(containerID string, networkID string, force bool) error {
+	nw, err := b.Daemon.FindNetwork(networkID)
+	if err != nil {
+		return err
+	}
+	return b.Daemon.DisconnectContainerFromNetwork(containerID, nw, force)
+}
+
+// DeleteNetwork destroys a network unless it's one of docker's predefined networks.
+func (b *Backend) DeleteNetwork(networkID string) error {
+	if _, err := b.clusterProvider.GetNetwork(networkID); err == nil {
+		return b.clusterProvider.RemoveNetwork(networkID)
+	}
+	if err := b.Daemon.DeleteNetwork(networkID); err != nil {
+		return err
+	}
+	return nil
+}
+
+func (b *Backend) buildNetworkResource(nw libnetwork.Network) *types.NetworkResource {
+	r := &types.NetworkResource{}
+	if nw == nil {
+		return r
+	}
+
+	info := nw.Info()
+	r.Name = nw.Name()
+	r.ID = nw.ID()
+	r.Scope = info.Scope()
+	if b.clusterProvider.IsManager() {
+		if _, err := b.clusterProvider.GetNetwork(nw.Name()); err == nil {
+			r.Scope = "swarm"
+		}
+	} else if info.Dynamic() {
+		r.Scope = "swarm"
+	}
+	r.Driver = nw.Type()
+	r.EnableIPv6 = info.IPv6Enabled()
+	r.Internal = info.Internal()
+	r.Options = info.DriverOptions()
+	r.Containers = make(map[string]types.EndpointResource)
+	buildIpamResources(r, info)
+	r.Internal = info.Internal()
+	r.Labels = info.Labels()
+
+	epl := nw.Endpoints()
+	for _, e := range epl {
+		ei := e.Info()
+		if ei == nil {
+			continue
+		}
+		sb := ei.Sandbox()
+		key := "ep-" + e.ID()
+		if sb != nil {
+			key = sb.ContainerID()
+		}
+
+		r.Containers[key] = buildEndpointResource(e)
+	}
+	return r
+}
+
+func buildIpamResources(r *types.NetworkResource, nwInfo libnetwork.NetworkInfo) {
+	id, opts, ipv4conf, ipv6conf := nwInfo.IpamConfig()
+
+	ipv4Info, ipv6Info := nwInfo.IpamInfo()
+
+	r.IPAM.Driver = id
+
+	r.IPAM.Options = opts
+
+	r.IPAM.Config = []network.IPAMConfig{}
+	for _, ip4 := range ipv4conf {
+		if ip4.PreferredPool == "" {
+			continue
+		}
+		iData := network.IPAMConfig{}
+		iData.Subnet = ip4.PreferredPool
+		iData.IPRange = ip4.SubPool
+		iData.Gateway = ip4.Gateway
+		iData.AuxAddress = ip4.AuxAddresses
+		r.IPAM.Config = append(r.IPAM.Config, iData)
+	}
+
+	if len(r.IPAM.Config) == 0 {
+		for _, ip4Info := range ipv4Info {
+			iData := network.IPAMConfig{}
+			iData.Subnet = ip4Info.IPAMData.Pool.String()
+			iData.Gateway = ip4Info.IPAMData.Gateway.String()
+			r.IPAM.Config = append(r.IPAM.Config, iData)
+		}
+	}
+
+	hasIpv6Conf := false
+	for _, ip6 := range ipv6conf {
+		if ip6.PreferredPool == "" {
+			continue
+		}
+		hasIpv6Conf = true
+		iData := network.IPAMConfig{}
+		iData.Subnet = ip6.PreferredPool
+		iData.IPRange = ip6.SubPool
+		iData.Gateway = ip6.Gateway
+		iData.AuxAddress = ip6.AuxAddresses
+		r.IPAM.Config = append(r.IPAM.Config, iData)
+	}
+
+	if !hasIpv6Conf {
+		for _, ip6Info := range ipv6Info {
+			iData := network.IPAMConfig{}
+			iData.Subnet = ip6Info.IPAMData.Pool.String()
+			iData.Gateway = ip6Info.IPAMData.Gateway.String()
+			r.IPAM.Config = append(r.IPAM.Config, iData)
+		}
+	}
+}
+
+func buildEndpointResource(e libnetwork.Endpoint) types.EndpointResource {
+	er := types.EndpointResource{}
+	if e == nil {
+		return er
+	}
+
+	er.EndpointID = e.ID()
+	er.Name = e.Name()
+	ei := e.Info()
+	if ei == nil {
+		return er
+	}
+
+	if iface := ei.Iface(); iface != nil {
+		if mac := iface.MacAddress(); mac != nil {
+			er.MacAddress = mac.String()
+		}
+		if ip := iface.Address(); ip != nil && len(ip.IP) > 0 {
+			er.IPv4Address = ip.String()
+		}
+
+		if ipv6 := iface.AddressIPv6(); ipv6 != nil && len(ipv6.IP) > 0 {
+			er.IPv6Address = ipv6.String()
+		}
+	}
+	return er
+}
