+++ api/server/router/system/system.go
@@ -2,22 +2,19 @@ package system
 
 import (
 	"github.com/docker/docker/api/server/router"
-	"github.com/docker/docker/daemon/cluster"
 )
 
 // systemRouter provides information about the Docker system overall.
 // It gathers information about host, daemon and container events.
 type systemRouter struct {
-	backend         Backend
-	clusterProvider *cluster.Cluster
-	routes          []router.Route
+	backend Backend
+	routes  []router.Route
 }
 
 // NewRouter initializes a new system router
-func NewRouter(b Backend, c *cluster.Cluster) router.Router {
+func NewRouter(b Backend) router.Router {
 	r := &systemRouter{
-		backend:         b,
-		clusterProvider: c,
+		backend: b,
 	}
 
 	r.routes = []router.Route{
